requirements.txt
aiohttp==3.7.4
astroid==2.5.1
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord==1.0.1
discord.py==1.6.0
idna==3.1
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
multidict==5.1.0
pylint==2.7.2
python-dotenv==0.15.0
PyYAML==6.0.2
toml==0.10.2
typing-extensions==3.7.4.3
wrapt==1.12.1
yarl==1.9.4
